requirements.txt
fastapi>=0.115.12
uvicorn[standard]==0.34.2
click>=8.0.0
colorama>=0.4.4
h11>=0.8
httptools>=0.4.0
python-dotenv>=0.13
pyyaml>=5.1
watchfiles>=0.13
websockets>=10.4
pillow==10.4.0
numpy==1.26.4
psycopg2-binary>=2.8.6
python-multipart>=0.0.5
requests>=2.31.0
